# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "S14S Identify - Enterprise Identifier Registry",
    "version": "1.0.0",
    "description": "API for managing enterprise identifiers with probabilistic matching and full audit trails."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/match-quality/feedback": {
      "get": {
        "summary": "List all match feedback records",
        "description": "Returns all match feedback records, optionally filtered by type.",
        "tags": [
          "Match Quality"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": ["false_positive", "false_negative"]
            },
            "description": "Filter by feedback type"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of match feedback records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchFeedback"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MatchFeedback": {
        "type": "object",
        "description": "A feedback record indicating a false positive or false negative match outcome.",
        "properties": {
          "_id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["false_positive", "false_negative"],
            "description": "Whether the feedback reports a false positive (incorrect auto-match) or false negative (missed match, detected via manual merge)"
          },
          "customer_id": {
            "type": "string",
            "description": "The customer record involved"
          },
          "related_customer_id": {
            "type": "string",
            "nullable": true,
            "description": "For false negatives, the other customer that should have matched"
          },
          "alias_id": {
            "type": "string",
            "nullable": true,
            "description": "For false positives, the alias that was incorrectly matched"
          },
          "original_confidence": {
            "type": "number",
            "nullable": true,
            "description": "The match confidence score at the time of the original auto-match"
          },
          "original_algorithm": {
            "type": "string",
            "nullable": true,
            "description": "The matching algorithm used (e.g. fellegi-sunter)"
          },
          "reported_by": {
            "type": "string",
            "description": "User who reported the feedback"
          },
          "reported_at": {
            "type": "string",
            "format": "date-time"
          },
          "resolved": {
            "type": "boolean",
            "description": "Whether this feedback has been resolved"
          },
          "resolved_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Optional notes explaining the feedback"
          }
        }
      },
...
    }
  },
...
}